+++ Tests/HTMLStringTests/HTMLStringTests.swift
@@ -1,16 +1,16 @@
-/*
+/**
  * ==---------------------------------------------------------------------------------==
  *
  *  File            :   HTMLStringTests.swift
  *  Project         :   HTMLString
- *  Author          :   ALEXIS AUBRY RADANOVIC
+ *  Author          :   Alexis Aubry Radanovic
  *
  *  License         :   The MIT License (MIT)
  *
  * ==---------------------------------------------------------------------------------==
  *
  *	The MIT License (MIT)
- *	Copyright (c) 2016 ALEXIS AUBRY RADANOVIC
+ *	Copyright (c) 2016-2017 Alexis Aubry Radanovic
  *
  *	Permission is hereby granted, free of charge, to any person obtaining a copy of
  *	this software and associated documentation files (the "Software"), to deal in
@@ -44,52 +44,6 @@ class HTMLStringTests: XCTestCase {
     // MARK: - Escaping
 
     ///
-    /// Tests escaping a character for ASCII.
-    ///
-
-    func testCharacterASCIIEscape() {
-
-        let namedEscape = Character("&").escapingForASCII
-        XCTAssertEqual(namedEscape, "&amp;")
-
-        let namedDualEscape = Character("⪰̸").escapingForASCII
-        XCTAssertEqual(namedDualEscape, "&NotSucceedsEqual;")
-
-        let emojiEscape = Character("🙃").escapingForASCII
-        XCTAssertEqual(emojiEscape, "&#128579;")
-
-        let doubleEmojiEscape = Character("🇺🇸").escapingForASCII
-        XCTAssertEqual(doubleEmojiEscape, "&#127482;&#127480;")
-
-        let basicCharacterEscape = Character("A").escapingForASCII
-        XCTAssertEqual(basicCharacterEscape, "A")
-
-    }
-
-    ///
-    /// Tests escaping a character for Unicode.
-    ///
-
-    func testCharacterUnicodeEscape() {
-
-        let requiredEscape = Character("&").escapingForUnicode
-        XCTAssertEqual(requiredEscape, "&amp;")
-
-        let namedDualEscape = Character("⪰̸").escapingForUnicode
-        XCTAssertEqual(namedDualEscape, "⪰̸")
-
-        let emojiEscape = Character("🙃").escapingForUnicode
-        XCTAssertEqual(emojiEscape, "🙃")
-
-        let doubleEmojiEscape = Character("🇺🇸").escapingForUnicode
-        XCTAssertEqual(doubleEmojiEscape, "🇺🇸")
-
-        let basicCharacterEscape = Character("A").escapingForUnicode
-        XCTAssertEqual(basicCharacterEscape, "A")
-
-    }
-
-    ///
     /// Tests escaping a string for ASCII.
     ///
 
@@ -99,7 +53,7 @@ class HTMLStringTests: XCTestCase {
         XCTAssertEqual(namedEscape, "Fish &amp; Chips")
 
         let namedDualEscape = ("a ⪰̸ b").escapingForASCIIHTML
-        XCTAssertEqual(namedDualEscape, "a &NotSucceedsEqual; b")
+        XCTAssertEqual(namedDualEscape, "a &#10928;&#824; b")
 
         let emojiEscape = ("Hey 🙃").escapingForASCIIHTML
         XCTAssertEqual(emojiEscape, "Hey &#128579;")
@@ -167,6 +121,9 @@ class HTMLStringTests: XCTestCase {
         let twoSequences = "a &amp;&amp; b".unescapingFromHTML
         XCTAssertEqual(twoSequences, "a && b")
 
+        let doubleEmojiEscape = ("Going to the &#127482;&#127480; next June").unescapingFromHTML
+        XCTAssertEqual(doubleEmojiEscape, "Going to the 🇺🇸 next June")
+
     }
 
     // MARK: - Benchmark
@@ -202,18 +159,6 @@ class HTMLStringTests: XCTestCase {
 
         self.measure {
 
-            _ = Character("&").escapingForASCII
-            _ = Character("⪰̸").escapingForASCII
-            _ = Character("🙃").escapingForASCII
-            _ = Character("🇺🇸").escapingForASCII
-            _ = Character("A").escapingForASCII
-
-            _ = Character("&").escapingForUnicode
-            _ = Character("⪰̸").escapingForUnicode
-            _ = Character("🙃").escapingForUnicode
-            _ = Character("🇺🇸").escapingForUnicode
-            _ = Character("A").escapingForUnicode
-
             _ = ("Fish & Chips").escapingForASCIIHTML
             _ = ("a ⪰̸ b").escapingForASCIIHTML
             _ = ("Hey 🙃").escapingForASCIIHTML
@@ -232,10 +177,8 @@ class HTMLStringTests: XCTestCase {
 
 extension HTMLStringTests {
 
-    static var allTests : [(String, (HTMLStringTests) -> () throws -> Void)] {
+    static var allTests: [(String, (HTMLStringTests) -> () throws -> Void)] {
         return [
-            ("testCharacterASCIIEscape", testCharacterASCIIEscape),
-            ("testCharacterUnicodeEscape", testCharacterUnicodeEscape),
             ("testStringASCIIEscaping", testStringASCIIEscaping),
             ("testStringUnicodeEscaping", testStringUnicodeEscaping),
             ("testUnescaping", testUnescaping),